pyproject.toml
[tool.poetry]
name = "OrbitESC"
version = "0.1.0"
description = "Brushless motor controller development environment"
authors = ["Brandon Braun <brandonbraun653@protonmail.com>"]
license = "MIT"
readme = "README.md"
packages = [
    { include = "pyorbit", from = "./" }
]

[tool.poetry.dependencies]
python = "^3.10"
cobs = "^1.2.1"
grpcio = "1.47.0"
grpcio-tools = "1.47.0"
protobuf = "3.19.1"
python-can = "^4.2.2"
jupyter = "^1.0.0"
numpy = "^1.23.4"
pyserial = "^3.5"
pytest = "^7.1.2"
pyqt5 = "==5.15.2"
loguru = "^0.7.2"
nanopb = "^0.4.7"
pyqtgraph = "^0.13.3"
pyudev = "^0.24.1"
pytest-profiling = "^1.7.0"
mypy-protobuf = "3.1.0"
sip = "^6.8.0"
astropy = "^6.0.0"
h5py = "^3.10.0"
tomli = "^2.0.1"
iminuit = "^2.24.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["pyorbit/tests"]
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s"